pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Flask web application that allows users to analyze web pages for SEO effectiveness."
authors = ["Sergey R <topgoods687@gmail.com>"]
readme = "README.md"
packages = [{include = "page_analyzer"}]

[tool.poetry.dependencies]
python = "^3.10"
gunicorn = "^20.1.0"
flask = "^3.0.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"
validators = "^0.22.0"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
